Added a new DataAttribute for TableColumnDefinition. Added 2 new Enums to support the new DataAttribute. Moved the TableColumnModel (which should be renamed to TableColumnDefinition).
