+++ GCFoundation.Components/Enums/CellAlignment.cs
@@ -1,0 +1,23 @@
+﻿namespace GCFoundation.Components.Enums
+{
+    /// <summary>
+    /// Use <see cref="CellAlignment"/> to control how the content inside the column cells is positioned horizontally.
+    /// </summary>
+    public enum CellAlignment
+    {
+        /// <summary>
+        /// Aligns the item to the center of the cell.
+        /// </summary>
+        center,
+
+        /// <summary>
+        /// Aligns the item to the end (right, by default) of the cell.
+        /// </summary>
+        end,
+
+        /// <summary>
+        /// Aligns the item to the start (left, by default) of the cell.
+        /// </summary>
+        start
+    }
+}
