+++ GCFoundation.Components/Enums/SortDirection.cs
@@ -1,0 +1,18 @@
+﻿namespace GCFoundation.Components.Enums
+{
+    /// <summary>
+    /// Defines the sort direction for a table column.
+    /// </summary>
+    public enum SortDirection
+    {
+        /// <summary>
+        /// Indicates ascending sort order.
+        /// </summary>
+        Ascending,
+
+        /// <summary>
+        /// Indicates descending sort order.
+        /// </summary>
+        Descending
+    }
+}
